feat: add remove user validations

+++ src/2-business/useCases/removeUserUseCase.ts
@@ -3,7 +3,7 @@ import { injectable, inject } from 'inversify'
 import { IUserRepository, IUserRepositoryToken } from '../repositories/iUserRepository'
 import { left, right } from '../../4-framework/shared/either'
 import { IUseCase } from './iUseCase'
-import { UserNotFound, UserRemovalFailed } from '../module/errors/users'
+import { UserIdentityCannotBeValidated, UserNotFound, UserRemovalFailed } from '../module/errors/users'
 import { InputRemoveUserDto, OutputRemoveUserDto } from '../dto/userDto'
 
 @injectable()
@@ -12,10 +12,15 @@ export class RemoveUserUseCase implements IUseCase<InputRemoveUserDto, OutputRem
 
   async exec(input: InputRemoveUserDto): Promise<OutputRemoveUserDto> {
     try {
+
+      if (input.user_id !== input.user_context_id) {
+        return left(UserIdentityCannotBeValidated)
+      }
+
       const userResult = await this.userRepository.remove(input.user_id)
-      
+
       if (!userResult) return left (UserNotFound)
-      
+
       return right(userResult)
     } catch (error) {
       return left(UserRemovalFailed)
